0417
git-svn-id: svn://172.16.18.3/onk/RK29kitchen@189 96281bfe-b365-5d4e-a99c-6d5232fd3cb0

+++ plugins/80.makeUpdateImage.sh
@@ -73,6 +73,8 @@ makeUpdateMkInitRD(){
 	cd ..	
 	abootimg --create boot.img -k $zimage -r ${initrd}
 
+	mkkrnlimg -r "${zimage}" kernel.img
+
 	popd
 }
 
@@ -82,7 +84,12 @@ makeUpdateImage(){
 	then
 		makeUpdateParseBL
 	fi
-	afptool -pack . ${img}.tmp
+	afptool -pack . ${img}.tmp 2>>"${LOGFILE}"
+	if [ ! -f "${img}.tmp" ]
+	then
+		dialogLOG "Make update image error"
+		exit 1
+	fi
 	commonBackupFile ${img}
 	img_maker $BOOTLOADER ${img}.tmp ${img}
 	rm ${img}.tmp
